# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: DraftFly API — AI-powered reply automation for B2B sales agencies
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List all campaigns
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign mapping
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /campaigns/{id}/stats:
    get:
      operationId: getCampaignStats
      tags: [campaigns]
      summary: Per-campaign reply and draft statistics
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignStats"

  # ---------- DRAFTS ----------
...
components:
  schemas:
...
    Campaign:
      type: object
      required: [id, clientId, name, lemlistCampaignId, isActive, replyCount, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        personaId:
          type: ["integer", "null"]
        name:
          type: string
        lemlistCampaignId:
          type: string
        tone:
          type: ["string", "null"]
        replyRules:
          type: ["string", "null"]
        regionRules:
          type: ["string", "null"]
        isActive:
          type: boolean
        replyCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    CampaignInput:
      type: object
      required: [clientId, name, lemlistCampaignId]
      properties:
        clientId:
          type: integer
        personaId:
          type: integer
        name:
          type: string
        lemlistCampaignId:
          type: string
        tone:
          type: string
        replyRules:
          type: string
        regionRules:
          type: string
        isActive:
          type: boolean
          description: >-
            Whether the campaign drafts replies. Omit to have it decided by the
            plan's remaining allowance — active when a slot is free, inactive
            when the client is already at their limit. Passing true while at the
            limit is refused with 422 rather than silently downgraded.
    CampaignUpdate:
      type: object
      properties:
        personaId:
          type: integer
        name:
          type: string
        lemlistCampaignId:
          type: string
        tone:
          type: string
        replyRules:
          type: string
        regionRules:
          type: string
        isActive:
          type: boolean
    CampaignStats:
      type: object
      required: [campaignId, totalReplies, sent, edited, discarded, pending]
      properties:
        campaignId:
          type: integer
        totalReplies:
          type: integer
        sent:
          type: integer
        edited:
          type: integer
        discarded:
          type: integer
        pending:
          type: integer
        avgResponseTimeMs:
          type: ["integer", "null"]

    # ---- Draft ----
...